requirements.txt
wikipedia
sentence-transformers
faiss-cpu
cohere
numpy